Include relative path in filename field for uploads
Previously it only set the full (relative) path for directories, but
not for individual files, so at the backend it could not be determined
which directory a file belonged to. Now both directories and files
contain the relative path correctly.

The current solution is  workaround using a customAction, but this is
actually an issue that needs to be fixed by vue-upload-component
https://github.com/lian-yue/vue-upload-component/issues/426