Add support for local storage with adhoc file upload
If "localhost" service is configured with

      "localhost": {
        "allowUpload": true
      }

then users will be able to upload files to the server to describe.

In this case "Upload" is displayed instead of "My Computer" and a new random folder
is created and assigned to the user, where the user can upload files. The random
folder's name is appended to each URL of the UI so that it can be bookmarked and
opened later to continue editing.

The UX is not too smooth yet but the basic functionality is there.